a program that finds and prints the first 98 Fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,13 +9,13 @@
 int main(void)
 {
 int i;
-long int num1 = 1, num2 = 2, sum;
+long double num1 = 1, num2 = 2, sum;
 for (i = 0; i < 97; i++)
 {
 sum = num1 + num2;
 num1 = num2;
 num2 = sum;
-printf("%ld", sum);
+printf("%.0lf", sum);
 if (i != 96)
 printf(", ");
 }
